% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_code_sav.R
\docType{package}
\name{CDSMutSig}
\alias{CDSMutSig}
\title{CDSMutSig: Mapping Coding Sequences between DNA Mutational signatures and Amino Acid Changes}
\description{
A set of functions which use Ensembl and Bioconductor genome objects
to map between given protein sequences to corresponding coding DNA
sequences, extract single-amino acid variants (SNVs) and corresponding
nucleotide substitutions.
}
