% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_code_sav.R
\name{get32Contexts}
\alias{get32Contexts}
\title{Get 32 DNA motifs used in defining mutational signatures}
\usage{
get32Contexts()
}
\value{
A character vector containing the 32 DNA motifs defined with the wild-type nucleotide (either C or T) at which single-base substitution occurs, together with 1 nt 5' and 1 nt 3'
}
\description{
Used as a general function to obtain all 32 possibilities for counting, graphing etc.
}
