% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_code_sav.R
\name{get96Contexts}
\alias{get96Contexts}
\title{Get 96 mutational contexts used in defining mutational signatures}
\usage{
get96Contexts()
}
\value{
A character vector containing the 96 mutational contexts defined with the single-base substitution (C>A, C>T, C>G, T>A, T>C, T>G) together with 1 nt 5' and 1 nt 3'.
}
\description{
Used as a general function to obtain all 96 possibilities for counting, graphing etc.
}
