% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_code_sav.R
\name{getExonProtBoundary}
\alias{getExonProtBoundary}
\title{Get amino acid sequence ranges residing in each exon}
\usage{
getExonProtBoundary(GRangesObj)
}
\arguments{
\item{GRangesObj}{A \code{GRangeObj} containing boundaries of each exon in the coding sequence. Generated internally in \code{mapCodon}.}
}
\value{
A data.frame with six columns:
  \code{exon}: exon number
  \code{lengths}: length (in nucleotides) of the exon
  \code{prot_start}: amino acid position at the beginning of the exon
  \code{prot_end}: amino acid position at the end of the exon
  \code{carried_forfward}: the number of nucleotides from the previous exon which does not yield a complete codon.
  \code{accu_length}: running sum of length (in nucleotides) of the coding DNA sequence.
}
\description{
Given a GRanges object of exon boundaries, this function determines
the range of amino acids residing in each exon. Split AA will appear in both exons.
}
