% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_code_sav.R
\name{getMutCodonPos}
\alias{getMutCodonPos}
\title{Get the codon position mutated to yield a given missense substitution}
\usage{
getMutCodonPos(wt_c, mut_c)
}
\arguments{
\item{wt_c}{wild-type codon}

\item{mut_c}{mutant codon}
}
\value{
A integer with value of 1, 2 or 3, corresponding to the position within the codon which is mutated from \code{wt_c} to \code{mut_c}.
}
\description{
Get the codon position mutated to yield a given missense substitution
}
\examples{
getMutCodonPos( "ATG", "CTG" ) # return 1
getMutCodonPos( "ATG", "ACG" ) # return 2
getMutCodonPos( "ATG", "ATA" ) # return 3

}
