% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_code_sav.R
\name{getPossibleMissenseSNVs}
\alias{getPossibleMissenseSNVs}
\title{Get possible missense substitutions arising from single-nucleotide variants}
\usage{
getPossibleMissenseSNVs()
}
\value{
A data.frame with four columns:
  \code{MUT}: codon corresponding to mutant amino acid
  \code{WT}: codon corresponding to wild-type amino acid
  \code{WT_AA}: wild-type amino acid (one-letter code)
  \code{MUT_AA}: wild-type amino acid (one-letter code)
}
\description{
This function removes missense substitutions that are impossible to achieve
without accumulating multiple single-nucleotide variants, based on the standard
genetic code.
}
