% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_code_sav.R
\name{mapCodon}
\alias{mapCodon}
\title{Map codons for each amino acid for a given protein.}
\usage{
mapCodon(prot_id, ensembldbObj, genomeObj, seqLevelStyle = "NCBI")
}
\arguments{
\item{prot_id}{character, Ensembl protein ID}

\item{ensembldbObj}{Ensembl DB Object from the \code{EnsDb} package (see example)}

\item{genomeObj}{Bioconductor genome object from which coding sequences are extracted (see example)}

\item{seqLevelStyle}{Either "UCSC" or "NCBI". Denote conventions used in genome object to name chromosomes. Need changing to match the genome object. (Default: "NCBI")}
}
\value{
A data.frame with four columns:
  \code{AA_pos}: integer, amino acid position of the given Ensembl protein
  \code{AA}: amino acid (one-letter code)
  \code{codon}: codon corresponding to \code{AA}
  \code{exon}: the exon(s) in which the given \code{AA} can be found. Amino acids which span across exons are denoted as a semicolon-separated triplets of exon IDs, with each entry corresponding to each position of the codon.
}
\description{
Map codons for each amino acid for a given protein.
}
\examples{
require(EnsDb.Hsapiens.v75)
require(BSgenome.Hsapiens.UCSC.hg19)
id <- "ENSP00000233146"
codons <- mapCodon(id, EnsDb.Hsapiens.v75::EnsDb.Hsapiens.v75,
                   BSgenome.Hsapiens.UCSC.hg19::Hsapiens,
                   seqLevelStyle = "UCSC")
}
