% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_code_sav.R
\name{mapMut}
\alias{mapMut}
\title{Map mutations given a table of codons for each amino acid for a given protein.}
\usage{
mapMut(codonMap, possibleMuts)
}
\arguments{
\item{codonMap}{data.frame of codons for each amino acid obtained from \code{mapCodon} (see example)}

\item{possibleMuts}{data.frame of possible mutations arising froms single-nucleotide susbtitutions obtained from \code{getPossibleMissenseSNVs} (see example)}
}
\value{
A data.frame with six columns:
  \code{AA}: wild-type amino acid (one-letter code)
  \code{WT_codon}: codon corresponding to wild-type amino acid
  \code{AA_pos}: integer, amino acid position of the given Ensembl protein
  \code{exon}: the exon(s) in which the given \code{AA} can be found. Amino acids which span across exons are denoted as a semicolon-separated triplets of exon IDs, with each entry corresponding to each position of the codon.
  \code{MUT_codon}: codon corresponding to mutant amino acid
  \code{MUT_AA}: mutant amino acid (one-letter code)
}
\description{
Map mutations given a table of codons for each amino acid for a given protein.
}
\examples{
require(EnsDb.Hsapiens.v75)
require(BSgenome.Hsapiens.UCSC.hg19)
id <- "ENSP00000233146"
codons <- mapCodon(id, EnsDb.Hsapiens.v75::EnsDb.Hsapiens.v75,
                   BSgenome.Hsapiens.UCSC.hg19::Hsapiens,
                   seqLevelStyle = "UCSC")
muts <- mapMut(codons, getPossibleMissenseSNVs())
}
